requirements.txt
asgiref==3.8.1                                                                                                                                                        
beautifulsoup4==4.12.3
Django==4.2.16
django-bootstrap-v5==1.0.11
django-recaptcha==4.0.0
python-dotenv==1.0.1
soupsieve==2.6
sqlparse==0.5.3
tzdata==2025.1